requirements.txt
# Core
pandas>=2.2
numpy>=1.26

# Database
sqlalchemy>=2.0
psycopg2-binary>=2.9

# API / HTTP
requests>=2.31
python-dotenv>=1.0

# Sentiment analysis
nltk>=3.9

# UI
streamlit>=1.38

plotly>=5.22    # interactive charts

bottleneck>=1.3.8
